# [Discussion] Using BufferedStream