--- a/src/libraries/System.Private.CoreLib/src/System/IO/FileStreamCompletionSource.Win32.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/FileStreamCompletionSource.Win32.cs
@@ -9,8 +9,13 @@ using System.Threading.Tasks;
 
 namespace System.IO
 {
-    internal sealed partial class WindowsFileStreamStrategy : FileStreamStrategyBase
+    internal sealed partial class AsyncWindowsFileStreamStrategy : WindowsFileStreamStrategy
     {
+        /// <summary>
+        /// owned by BufferedStream, do NOT use outside of FileStreamCompletionSource
+        /// </summary>
+        private object? _buffer;
+
         // This is an internal object extending TaskCompletionSource with fields
         // for all of the relevant data necessary to complete the IO operation.
         // This is used by IOCallback and all of the async methods.
@@ -25,7 +30,7 @@ namespace System.IO
 
             private static Action<object?>? s_cancelCallback;
 
-            private readonly WindowsFileStreamStrategy _stream;
+            private readonly AsyncWindowsFileStreamStrategy _stream;
             private readonly int _numBufferedBytes;
             private CancellationTokenRegistration _cancellationRegistration;
 #if DEBUG
@@ -35,7 +40,7 @@ namespace System.IO
             private long _result; // Using long since this needs to be used in Interlocked APIs
 
             // Using RunContinuationsAsynchronously for compat reasons (old API used Task.Factory.StartNew for continuations)
-            protected FileStreamCompletionSource(WindowsFileStreamStrategy stream, int numBufferedBytes, byte[]? bytes)
+            protected FileStreamCompletionSource(AsyncWindowsFileStreamStrategy stream, int numBufferedBytes, byte[]? bytes)
                 : base(TaskCreationOptions.RunContinuationsAsynchronously)
             {
                 _numBufferedBytes = numBufferedBytes;
@@ -132,8 +137,8 @@ namespace System.IO
                 // be directly the FileStreamCompletionSource that's completing (in the case where the preallocated
                 // overlapped was already in use by another operation).
                 object? state = ThreadPoolBoundHandle.GetNativeOverlappedState(pOverlapped);
-                Debug.Assert(state is WindowsFileStreamStrategy || state is FileStreamCompletionSource);
-                FileStreamCompletionSource completionSource = state is WindowsFileStreamStrategy fs ?
+                Debug.Assert(state is AsyncWindowsFileStreamStrategy || state is FileStreamCompletionSource);
+                FileStreamCompletionSource completionSource = state is AsyncWindowsFileStreamStrategy fs ?
                     fs._currentOverlappedOwner! : // must be owned
                     (FileStreamCompletionSource)state!;
                 Debug.Assert(completionSource != null);
@@ -220,7 +225,7 @@ namespace System.IO
                 }
             }
 
-            public static FileStreamCompletionSource Create(WindowsFileStreamStrategy stream, int numBufferedBytesRead, ReadOnlyMemory<byte> memory)
+            public static FileStreamCompletionSource Create(AsyncWindowsFileStreamStrategy stream, int numBufferedBytesRead, ReadOnlyMemory<byte> memory)
             {
                 // If the memory passed in is the stream's internal buffer, we can use the base FileStreamCompletionSource,
                 // which has a PreAllocatedOverlapped with the memory already pinned.  Otherwise, we use the derived
@@ -241,7 +246,7 @@ namespace System.IO
         {
             private MemoryHandle _handle; // mutable struct; do not make this readonly
 
-            internal MemoryFileStreamCompletionSource(WindowsFileStreamStrategy stream, int numBufferedBytes, ReadOnlyMemory<byte> memory) :
+            internal MemoryFileStreamCompletionSource(AsyncWindowsFileStreamStrategy stream, int numBufferedBytes, ReadOnlyMemory<byte> memory) :
                 base(stream, numBufferedBytes, bytes: null) // this type handles the pinning, so null is passed for bytes
             {
                 _handle = memory.Pin();