… src/libraries/System.IO.FileSystem/tests/FileStream/Flush.cs
… src/libraries/System.IO.FileSystem/tests/FileStream/FlushAsync.cs
… src/libraries/System.IO.FileSystem/tests/FileStream/SafeFileHandle.cs
… src/libraries/System.IO.FileSystem/tests/FileStream/WriteAsync.cs
… src/libraries/System.Private.CoreLib/src/System/IO/BufferedStream.cs
… src/libraries/System.Private.CoreLib/src/System/IO/WindowsFileStreamStrategy.cs
… src/libraries/System.Private.CoreLib/src/System/Threading/SemaphoreSlim.cs