@carlossanlop @Jozkee @stephentoub Please note that this PR is not even a draft, it's a PR in my fork. I've opened it as I want to get your opinion before I invest more time in it.

For the last couple of days I have been trying to wrap my head around:

* Win32 FileStream turns async reads into sync reads https://github.com/dotnet/runtime/issues/16341
* FileStream.FlushAsync ends up doing synchronous writes https://github.com/dotnet/runtime/issues/27643

@stephentoub has suggested [here](https://github.com/dotnet/runtime/issues/16341#issuecomment-293347286) how this could be solved.
To be honest with you I don't like this approach as in my opinion it would not be 100% thread-safe (it's OK, FileStream was never promised to be thread-safe) and it would make FileStream **even more complex**. 
It's definitely doable but if it's possible I would prefer to avoid making FileStream even more complex.

Stephen has also suggested using `BufferedStream` and removing all the buffering logic from the strategies themselves: https://github.com/dotnet/runtime/pull/47128#discussion_r560395799
I love the idea of removing the buffering logic from `FileStream`. And this is what I've tried in this PR (branch). It makes both `AsyncWindowsFileStreamStrategy` and `SyncWindowsFileStreamStrategy` super simple and very clean (no if checks, just tiny wrappers around OS calls, see the code below). `SyncWindowsFileStreamStrategy` is just 169 lines of code (half of it are comments)!!

Moreover, `BufferedStream` **solves the problem of doing sync method calls from async methods to handle the buffering**.

`BufferedStream` uses `SempahoreSlim` to achieve that:

https://github.com/dotnet/runtime/blob/16a9d5d6a2290a4f13fc4b42bf433588d481c39f/src/libraries/System.Private.CoreLib/src/System/IO/BufferedStream.cs#L661-L663

This comes with a performance penalty but also with a breaking change.

Until now, it was possible to start multiple async reads|writes and await them later. By using `BufferedStream` that acquires a lock it's still possible, but the requests would not be executed in parallel and `FileStream.Position` would get updated **after the request would be finished**:

https://github.com/dotnet/runtime/blob/16a9d5d6a2290a4f13fc4b42bf433588d481c39f/src/libraries/System.Private.CoreLib/src/System/IO/BufferedStream.cs#L1118-L1119

Test that fails after my changes:

https://github.com/dotnet/runtime/blob/16a9d5d6a2290a4f13fc4b42bf433588d481c39f/src/libraries/System.IO.FileSystem/tests/FileStream/WriteAsync.cs#L218-L229

If I remember correctly, we were initially OK with breaking that. But, if our users really want to perform async reads and writes in parallel, we could add offset-based methods to `FileStream` with a new mode (like `FileOptions.Overlapped`) that would not be using buffering at all (https://github.com/dotnet/runtime/issues/24847):

```diff
public partial class FileStream : System.IO.Stream
{
    public override int Read(byte[] buffer, int offset, int count)
    public override int Read(System.Span<byte> buffer)
+   public virtual int Read(System.Span<byte> buffer, long fileOffset)
    public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
    public override ValueTask<int> ReadAsync(System.Memory<byte> buffer, CancellationToken cancellationToken)
+   public virtual ValueTask<int> ReadAsync(System.Memory<byte> buffer, long fileOffset, CancellationToken cancellationToken)
    public override void Write(byte[] buffer, int offset, int count) { }
    public override void Write(ReadOnlySpan<byte> buffer) { }
+   public virtual void Write(ReadOnlySpan<byte> buffer, long fileOffset) { }    
    public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
    public override ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken)
+   public virtual ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, long fileOffset, CancellationToken cancellationToken)
}
```

So on the one hand we would break 0.001% of our users that does that, but on the other, we would offer them a super-fast alternative (this is what most probably they wanted from the beginning).

Speaking of the performance penalty I've spent some time tunning `BufferedStream` perf and I was able to get to the point that there is no perf regression in most of the cases.

<details>

`\before\CoreRun.exe` is 6.0 without the concept of `FileStreamStrategy`:


|          Method |        Job |           Toolchain |  fileSize | userBufferSize |      options |          Median | Ratio |    Allocated |
|---------------- |----------- |-------------------- |---------- |--------------- |------------- |----------------:|------:|-------------:|
|            Read | Job-RQDFGV |  \after\CoreRun.exe |      1024 |            512 |         None |        50.33 us |  0.99 |      4,384 B |
|            Read | Job-TPLJRT | \before\CoreRun.exe |      1024 |            512 |         None |        50.88 us |  1.00 |      4,288 B |
|                 |            |                     |           |                |              |                 |       |              |
|           Write | Job-RQDFGV |  \after\CoreRun.exe |      1024 |            512 |         None |       895.70 us |  1.00 |      4,384 B |
|           Write | Job-TPLJRT | \before\CoreRun.exe |      1024 |            512 |         None |       899.52 us |  1.00 |      4,288 B |
|                 |            |                     |           |                |              |                 |       |              |
|       ReadAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |            512 |         None |        69.46 us |  1.00 |      5,128 B |
|       ReadAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |            512 |         None |        69.71 us |  1.00 |      4,848 B |
|                 |            |                     |           |                |              |                 |       |              |
|      WriteAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |            512 |         None |       911.77 us |  0.98 |      4,472 B |
|      WriteAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |            512 |         None |       928.22 us |  1.00 |      4,840 B |
|                 |            |                     |           |                |              |                 |       |              |
|       ReadAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |            512 | Asynchronous |        86.91 us |  1.25 |      5,304 B |
|       ReadAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |            512 | Asynchronous |        69.49 us |  1.00 |      4,713 B |
|                 |            |                     |           |                |              |                 |       |              |
|      WriteAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |            512 | Asynchronous |       985.40 us |  1.04 |      4,900 B |
|      WriteAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |            512 | Asynchronous |       938.65 us |  1.00 |      4,693 B |
|                 |            |                     |           |                |              |                 |       |              |
|            Read | Job-RQDFGV |  \after\CoreRun.exe |      1024 |           4096 |         None |        49.06 us |  1.00 |        264 B |
|            Read | Job-TPLJRT | \before\CoreRun.exe |      1024 |           4096 |         None |        49.19 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|           Write | Job-RQDFGV |  \after\CoreRun.exe |      1024 |           4096 |         None |       181.07 us |  1.00 |        264 B |
|           Write | Job-TPLJRT | \before\CoreRun.exe |      1024 |           4096 |         None |       181.63 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|       ReadAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |           4096 |         None |        66.77 us |  1.03 |        800 B |
|       ReadAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |           4096 |         None |        64.61 us |  1.00 |        616 B |
|                 |            |                     |           |                |              |                 |       |              |
|      WriteAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |           4096 |         None |       182.16 us |  1.00 |        264 B |
|      WriteAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |           4096 |         None |       181.49 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|       ReadAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |           4096 | Asynchronous |        85.35 us |  1.03 |        968 B |
|       ReadAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |           4096 | Asynchronous |        82.89 us |  1.00 |        760 B |
|                 |            |                     |           |                |              |                 |       |              |
|      WriteAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |           4096 | Asynchronous |       183.62 us |  1.00 |        320 B |
|      WriteAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |           4096 | Asynchronous |       184.87 us |  1.00 |        200 B |
|                 |            |                     |           |                |              |                 |       |              |
|       OpenClose | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |        41.90 us |  1.01 |        264 B |
|       OpenClose | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |        41.68 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|      LockUnlock | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |        81.99 us |  1.04 |        264 B |
|      LockUnlock | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |        78.81 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|     SeekForward | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |       269.15 us |  1.04 |        264 B |
|     SeekForward | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |       258.58 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|    SeekBackward | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |     1,660.26 us |  1.01 |        265 B |
|    SeekBackward | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |     1,661.86 us |  1.00 |        168 B |
|                 |            |                     |           |                |              |                 |       |              |
|        ReadByte | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |        52.67 us |  0.96 |      4,384 B |
|        ReadByte | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |        54.19 us |  1.00 |      4,288 B |
|                 |            |                     |           |                |              |                 |       |              |
|       WriteByte | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |       927.91 us |  1.01 |      4,385 B |
|       WriteByte | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |       931.51 us |  1.00 |      4,288 B |
|                 |            |                     |           |                |              |                 |       |              |
|           Flush | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |     3,759.49 us |  1.00 |      4,386 B |
|           Flush | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |     3,802.99 us |  1.00 |      4,289 B |
|                 |            |                     |           |                |              |                 |       |              |
|      FlushAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |     7,357.75 us |  1.95 |    496,291 B |
|      FlushAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |     3,776.47 us |  1.00 |      4,306 B |
|                 |            |                     |           |                |              |                 |       |              |
|      CopyToFile | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |       986.10 us |  1.00 |      8,770 B |
|      CopyToFile | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |       998.14 us |  1.00 |      8,577 B |
|                 |            |                     |           |                |              |                 |       |              |
| CopyToFileAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? |         None |       985.46 us |  1.01 |      5,674 B |
| CopyToFileAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? |         None |       983.22 us |  1.00 |      9,465 B |
|                 |            |                     |           |                |              |                 |       |              |
|       OpenClose | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |        43.52 us |  0.94 |        320 B |
|       OpenClose | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |        46.30 us |  1.00 |        200 B |
|                 |            |                     |           |                |              |                 |       |              |
|      LockUnlock | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |        86.53 us |  1.00 |        320 B |
|      LockUnlock | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |        88.61 us |  1.00 |        200 B |
|                 |            |                     |           |                |              |                 |       |              |
|     SeekForward | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |     2,391.24 us |  0.99 |        321 B |
|     SeekForward | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |     2,406.04 us |  1.00 |        201 B |
|                 |            |                     |           |                |              |                 |       |              |
|    SeekBackward | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |     4,811.03 us |  0.98 |        322 B |
|    SeekBackward | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |     4,843.99 us |  1.00 |        201 B |
|                 |            |                     |           |                |              |                 |       |              |
|        ReadByte | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |        71.92 us |  0.93 |      4,760 B |
|        ReadByte | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |        76.48 us |  1.00 |      4,641 B |
|                 |            |                     |           |                |              |                 |       |              |
|       WriteByte | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |     1,004.14 us |  1.04 |      4,805 B |
|       WriteByte | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |       966.91 us |  1.00 |      4,670 B |
|                 |            |                     |           |                |              |                 |       |              |
|           Flush | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |    29,860.54 us |  1.68 |    152,195 B |
|           Flush | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |    17,744.32 us |  1.00 |    152,063 B |
|                 |            |                     |           |                |              |                 |       |              |
|      FlushAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |    35,450.52 us |  2.05 |    529,224 B |
|      FlushAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |    17,602.27 us |  1.00 |    152,279 B |
|                 |            |                     |           |                |              |                 |       |              |
| CopyToFileAsync | Job-RQDFGV |  \after\CoreRun.exe |      1024 |              ? | Asynchronous |     1,095.60 us |  1.13 |      6,122 B |
| CopyToFileAsync | Job-TPLJRT | \before\CoreRun.exe |      1024 |              ? | Asynchronous |     1,060.03 us |  1.00 |      9,860 B |
|                 |            |                     |           |                |              |                 |       |              |
|            Read | Job-RQDFGV |  \after\CoreRun.exe |   1048576 |            512 |         None |       711.89 us |  1.01 |      4,384 B |
|            Read | Job-TPLJRT | \before\CoreRun.exe |   1048576 |            512 |         None |       709.69 us |  1.00 |      4,288 B |
|                 |            |                     |           |                |              |                 |       |              |
|           Write | Job-RQDFGV |  \after\CoreRun.exe |   1048576 |            512 |         None |     6,203.04 us |  1.00 |      4,387 B |
|           Write | Job-TPLJRT | \before\CoreRun.exe |   1048576 |            512 |         None |     6,442.30 us |  1.00 |      4,289 B |
|                 |            |                     |           |                |              |                 |       |              |
|       ReadAsync | Job-RQDFGV |  \after\CoreRun.exe |   1048576 |            512 |         None |     1,628.45 us |  0.46 |     86,729 B |
|       ReadAsync | Job-TPLJRT | \before\CoreRun.exe |   1048576 |            512 |         None |     3,509.77 us |  1.00 |    234,002 B |
|                 |            |                     |           |                |              |                 |       |              |
|      WriteAsync | Job-RQDFGV |  \after\CoreRun.exe |   1048576 |            512 |         None |     8,688.55 us |  0.85 |     78,534 B |
|      WriteAsync | Job-TPLJRT | \before\CoreRun.exe |   1048576 |            512 |         None |    10,451.96 us |  1.00 |    233,995 B |
|                 |            |                     |           |                |              |                 |       |              |
|       ReadAsync | Job-RQDFGV |  \after\CoreRun.exe |   1048576 |            512 | Asynchronous |     4,675.79 us |  1.18 |     95,063 B |
|       ReadAsync | Job-TPLJRT | \before\CoreRun.exe |   1048576 |            512 | Asynchronous |     3,979.44 us |  1.00 |     41,577 B |
…